Add solution to exercise 4.21

+++ part4/bloglist/controllers/blogsController.js
@@ -26,6 +26,17 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== decodedToken.id) {
+    return response.status(401).json({ error: 'unauthorized action' })
+  }
+  const user = await User.findById(blog.user.toString())
+  user.blogs = user.blogs.filter((blog) => blog.toString() !== request.params.id)
+  await user.save()
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
